Handle the armed status better
Blink allows users to arm/disarm individual cameras. This is pretty useful for
inside and outside cameras, for instance.

The 'armed' information in the JSON however, merely indicates whether the
*system* is armed, not whether the camera itself is armed. This value, which is
the same as blink.arm, does not indicate whether the camera has motion
detection enabled.

In the JSON, the true status information lies in the 'active' key. The value
can be 'armed', 'disarmed' or 'disabled'. This commit reflects that value
in the armed property.